ISSUE #1 | Restructure project and modify build process.
