+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import logo from '../../logo.svg';
-import './App.css';
+import logo from './logo.svg';
+import './App.scss';
 
 const App: React.FC = () => {
   return (
