+++ nodemon.json
@@ -1,6 +1,6 @@
 {
-    "watch": ["src"],
-    "ext": ".ts,.js",
+    "watch": ["server", "client/src"],
+    "ext": ".ts,.tsx,.js",
     "ignore": [],
-    "exec": "ts-node ./src/server.ts"
+    "exec": "ts-node ./build/server.js"
   }
